Source/WebCore: [CSS Regions][CSS Shapes] Layout error when the shape has negative top coordinate and it's applied on the second region <https://webkit.org/b/123346>
Reviewed by David Hyatt.

We have a layout error when there is a shape applied on the second region, and it has a negative 'top' coordinate.
Since shapeInsideInfo::shapeLogicalTop() can return negative numbers, we need to check for it, when we're positioning
the first line in the region.

Test: fast/regions/shape-inside/shape-inside-on-multiple-regions-with-negative-shape-top.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlockFlow::updateShapeAndSegmentsForCurrentLineInFlowThread):

LayoutTests: [CSS Regions][CSS Shapes] Layout error when the shape has negative top coordinate and it's applied on the second region
<http://webkit.org/b/123346>

Reviewed by David Hyatt.

We have a layout error when there is a shape applied on the second region, and it has a negative 'top' coordinate.
Since shapeInsideInfo::shapeLogicalTop() can return negative numbers, we need to check for it, when we're positioning
the first line in the region.

* fast/regions/shape-inside/shape-inside-on-multiple-regions-with-negative-shape-top-expected.html: Added.
* fast/regions/shape-inside/shape-inside-on-multiple-regions-with-negative-shape-top.html: Added.

Identifier: 141589@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158219 268f45cc-cd09-0410-ab3c-d52691b4dbfc