It should be possible to specify extra tests to run-javascriptcore-tests https://bugs.webkit.org/show_bug.cgi?id=123495
Reviewed by Oliver Hunt.

This adds two ways to specify additional tests: the option --extra-tests=<blah>, and
the environment variable EXTRA_JSC_TESTS.

Those two values (the command-line option and the environment variable) are passed
as extra arguments to run-jsc-stress-tests.

To make this extra super useful, run-jsc-stress-tests now allows yaml configuration
files to be recursive: if a "- collection:" line is found, the string is handled
just as if it was a command-line argument. This allows you to use --extra-tests=file
where file contains a bunch of "- collection:" lines, and this is equivalent to each
of those lines being a command-line argument.

* Scripts/run-javascriptcore-tests:
* Scripts/run-jsc-stress-tests:



Identifier: 141623@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158255 268f45cc-cd09-0410-ab3c-d52691b4dbfc