Generated color wheel displays incorrectly (regressed in r155567) https://bugs.webkit.org/show_bug.cgi?id=123664
Source/JavaScriptCore:

Reviewed by Andreas Kling.

Interestingly, r155567 just "un-broke" the attempt to constant-fold ArithMod, but
that constant folding was just wrong to begin with. There is no evidence that this
constant folding rule is profitable. I'm removing it instead of trying to think
about what it means for it to be correct.

* dfg/DFGAbstractInterpreterInlines.h:
(JSC::DFG::::executeEffects):

LayoutTests:

Reviewed by Andreas Kling.

* js/dfg-mod-1-expected.txt: Added.
* js/dfg-mod-1.html: Added.
* js/script-tests/dfg-mod-1.js: Added.
(foo):



Identifier: 141909@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158556 268f45cc-cd09-0410-ab3c-d52691b4dbfc