[CSS Shapes] image valued shape element margin can cause an ASSERT fail https://bugs.webkit.org/show_bug.cgi?id=123743
Reviewed by Andreas Kling.

Source/WebCore:

When margin-top is specified for a shape's element, the Y coordinates passed
to RasterShapeIntervals::getExcludedIntervals() may be negative. This is
because the incoming logical coordinates are relative to the content or
padding box, depending on the box-sizing property. The RasterShape::getExcludedIntervals()
method now clips the incoming Y coordinates to the shape's bounds before
accessing the shape's intervals.

Test: fast/shapes/shape-outside-floats/shape-outside-floats-margin-crash.html

* rendering/shapes/RasterShape.cpp:
(WebCore::RasterShapeIntervals::getExcludedIntervals):

LayoutTests:

Verify that an element with an image valued shape and a non-negative margin
does not cause an assertion failure.

* fast/shapes/shape-outside-floats/shape-outside-floats-margin-crash-expected.html: Added.
* fast/shapes/shape-outside-floats/shape-outside-floats-margin-crash.html: Added.


Identifier: 141942@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158596 268f45cc-cd09-0410-ab3c-d52691b4dbfc