SMIL timers can still fire after the containing document has been torn down <https://webkit.org/b/123291>
Reviewed by Darin Adler.

Source/WebCore:

Ensure that the timers get paused when the document is preparing to be torn down.

Test: svg/animations/smil-timers-not-disabled-crash.html

* dom/Document.cpp:
(WebCore::Document::dropChildren):
(WebCore::Document::commonTeardown):
(WebCore::Document::prepareForDestruction):
* dom/Document.h:

LayoutTests:

Add a test that triggers the crash; unfortunately it triggers it after
the test is done, causing the next test to time out or crash.

* svg/animations/smil-timers-not-disabled-crash-expected.txt: Added.
* svg/animations/smil-timers-not-disabled-crash.html: Added.


Identifier: 141961@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158627 268f45cc-cd09-0410-ab3c-d52691b4dbfc