DumpRenderTree should reset its preferences file on launch https://bugs.webkit.org/show_bug.cgi?id=29751
Reviewed by Dan Bernstein.

Source/WebKit/mac:

* WebView/WebView.mm:
(-[WebView _retrieveKeyboardUIModeFromPreferences:]):
Further fix reading of AppleKeyboardUIMode - reading it from kCFPreferencesAnyApplication
domain meant that current application's domain was explcitly ignored, and the only
way to override the setting in DRT was through argument domain.

Source/WebKit2:

* UIProcess/mac/WKFullKeyboardAccessWatcher.mm:
(-[WKFullKeyboardAccessWatcher retrieveKeyboardUIModeFromPreferences:]):
Further fix reading of AppleKeyboardUIMode - reading it from kCFPreferencesAnyApplication
domain meant that current application's domain was explcitly ignored, and the only
way to override the setting in WTR was through argument domain.

Tools:

Using NSArgumentDomain was not a good idea. While it works with NSUserDefaults
and CFPreferencesCopyAppValue family of functions, it doesn't work with
CFPreferencesCopyValue(..., kCFPreferencesCurrentApplication, ...).

So, now we'll just set defaults in the current application.

* DumpRenderTree/mac/DumpRenderTree.mm: (setDefaultsToConsistentValuesForTesting):
- Delete the defaults.
- Removed CFPreferencesSetAppValue calls, as -setValuesForKeysWithDictionary: does the same.
- The comment about [NSUserDefaults standardUserDefaults] indirectly initializing NSApplication was incorrect.

* WebKitTestRunner/mac/main.mm: (setDefaultsToConsistentValuesForTesting):
Switched to -setValuesForKeysWithDictionary for consistency.



Identifier: 141979@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158652 268f45cc-cd09-0410-ab3c-d52691b4dbfc